+++ SimpleGUICS2Pygame/simplegui_lib.py
@@ -1,7 +1,7 @@
 # -*- coding: latin-1 -*-
 
 """
-simplegui_lib (May 31, 2015)
+simplegui_lib (November 21, 2015)
 
 Some functions and classes to help
 in SimpleGUI of CodeSkulptor,
@@ -18,12 +18,14 @@ http://www.opimedia.be/
 """
 
 try:
-    from user23_HY71NDvHu7WKaMa import draw_rect, draw_text_side
+    from user40_GjCFdIeSPOViuUZ \
+        import draw_rect, draw_text_multi, draw_text_side
     from user33_Bhc7VzXKbXGVQV1 import FPS
     from user30_VE5zdyz5OZwgen9 import Keys
     from user40_nMs7JxzimyImAv2 import Loader
 except ImportError:
-    from SimpleGUICS2Pygame.simplegui_lib_draw import draw_rect, draw_text_side
+    from SimpleGUICS2Pygame.simplegui_lib_draw \
+        import draw_rect, draw_text_multi, draw_text_side
     from SimpleGUICS2Pygame.simplegui_lib_fps import FPS
     from SimpleGUICS2Pygame.simplegui_lib_keys import Keys
     from SimpleGUICS2Pygame.simplegui_lib_loader import Loader