+++ Sphinx/ChangeLog.rst
@@ -1,6 +1,6 @@
 ChangeLog
 =========
-* 02.00.00 WORKING VERSION --- November 23, 2015
+* 02.00.00 WORKING VERSION --- January 6, 2016
 
   - Added ``script/pygame_check.py`` to check Pygame installation alone.